add guest management

+++ src/Controller/Admin/MediaController.php
@@ -2,18 +2,27 @@
 
 namespace App\Controller\Admin;
 
+use App\Entity\User;
 use App\Entity\Media;
 use App\Form\MediaType;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
+#[IsGranted('ROLE_USER')]
 class MediaController extends AbstractController
 {
-    /**
-     * @Route("/admin/media", name="admin_media_index")
-     */
-    public function index(Request $request)
+    public function __construct(
+        private EntityManagerInterface $entityManager
+    )
+    {
+    }
+
+    #[Route('/admin/media', name: 'admin_media_index')]
+    public function index(Request $request): Response
     {
         $page = $request->query->getInt('page', 1);
 
@@ -23,13 +32,13 @@ class MediaController extends AbstractController
             $criteria['user'] = $this->getUser();
         }
 
-        $medias = $this->getDoctrine()->getRepository(Media::class)->findBy(
+        $medias = $this->entityManager->getRepository(Media::class)->findBy(
             $criteria,
             ['id' => 'ASC'],
             25,
             25 * ($page - 1)
         );
-        $total = $this->getDoctrine()->getRepository(Media::class)->count([]);
+        $total = $this->entityManager->getRepository(Media::class)->count([]);
 
         return $this->render('admin/media/index.html.twig', [
             'medias' => $medias,
@@ -38,10 +47,8 @@ class MediaController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/admin/media/add", name="admin_media_add")
-     */
-    public function add(Request $request)
+    #[Route('/admin/media/add', name: 'admin_media_add')]
+    public function add(Request $request): Response
     {
         $media = new Media();
         $form = $this->createForm(MediaType::class, $media, ['is_admin' => $this->isGranted('ROLE_ADMIN')]);
@@ -49,29 +56,40 @@ class MediaController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             if (!$this->isGranted('ROLE_ADMIN')) {
-                $media->setUser($this->getUser());
-            }
-            $media->setPath('uploads/' . md5(uniqid()) . '.' . $media->getFile()->guessExtension());
-            $media->getFile()->move('uploads/', $media->getPath());
-            $this->getDoctrine()->getManager()->persist($media);
-            $this->getDoctrine()->getManager()->flush();
+                $user = $this->getUser();
 
+                if ($user instanceof User) {
+                    $media->setUser($user);
+                }
+            }
+
+            $media->setPath('uploads/' . md5(uniqid('', true)) . '.' . $media->getFile()->guessExtension());
+            $media->getFile()->move('uploads/', $media->getPath());
+
+            $this->entityManager->persist($media);
+            $this->entityManager->flush();
+            $this->addFlash('success', 'Média bien ajouté');
+            
             return $this->redirectToRoute('admin_media_index');
         }
 
         return $this->render('admin/media/add.html.twig', ['form' => $form->createView()]);
     }
 
-    /**
-     * @Route("/admin/media/delete/{id}", name="admin_media_delete")
-     */
-    public function delete(int $id)
+    #[Route('/admin/media/delete/{id}', name: 'admin_media_delete')]    
+    public function delete(int $id): Response
     {
-        $media = $this->getDoctrine()->getRepository(Media::class)->find($id);
-        $this->getDoctrine()->getManager()->remove($media);
-        $this->getDoctrine()->getManager()->flush();
-        unlink($media->getPath());
+        $media = $this->entityManager->getRepository(Media::class)->find($id);
 
-        return $this->redirectToRoute('admin_media_index');
+        $user = $this->getUser();
+
+        if ($this->isGranted('ROLE_ADMIN') || ($media->getUser() === $user)) {
+            $this->entityManager->remove($media);
+            $this->entityManager->flush();
+
+            unlink($media->getPath());    
+        }
+
+        return $this->redirectToRoute('admin_media_index');    
     }
 }
